requirements.txt
confluent-kafka>=2.3.0
pytest>=8.3.0
jsonschema>=4.22.0
flask>=3.0.0
google-cloud-firestore>=2.16.0
google-cloud-aiplatform>=1.71.0
